third_party/blink/web_tests/external/wpt/css/css-values/tree-counting/sibling-function-descriptors.tentative.html

<!DOCTYPE html>
<title>CSS Values and Units Test: sibling-index() and sibling-count() invalid in descriptors</title>
<link rel="help" href="https://drafts.csswg.org/css-values-5/#tree-counting">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/10982">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style id="page_sheet">
  @page {
    margin: 100px;
    margin: calc(0px * sibling-index());
  }
  @page {
    margin: 100px;
    margin: calc(0px * sibling-count());
  }
</style>
<style id="font_face_sheet">
  @font-face {
    font-family: my-font;
    font-weight: 300;
    font-weight: calc(max(0 * sibling-index(), 400));
    font-feature-settings: "vert" 2;
    font-feature-settings: "vert" calc(max(sibling-index(), 1));
  }
  @font-face {
    font-family: my-font;
    font-weight: 300;
    font-weight: calc(max(0 * sibling-count(), 400));
    font-feature-settings: "vert" 2;
    font-feature-settings: "vert" calc(max(sibling-count(), 1));
  }
</style>
<style id="font_palette_sheet">
  @font-palette-values --foo {
    font-family: my-font;
    base-palette: 1;
    base-palette: calc(max(sibling-index(), 2));
    override-colors: 1 green;
    override-colors: sibling-index() red;
  }
  @font-palette-values --foo {
    font-family: my-font;
    base-palette: 1;
    base-palette: calc(max(sibling-count(), 2));
    override-colors: 1 green;
    override-colors: sibling-count() red;
  }
</style>
<script>
  const page_rules = page_sheet.sheet.cssRules;

  test(() => {
    assert_equals(page_rules[0].style.margin, "100px");
  }, "sibling-index() should not be allowed in @page properties");

  test(() => {
    assert_equals(page_rules[1].style.margin, "100px");
  }, "sibling-count() should not be allowed in @page properties");

  const font_face_rules = font_face_sheet.sheet.cssRules;

  test(() => {
    assert_equals(font_face_rules[0].style.fontWeight, "300");
    assert_equals(font_face_rules[0].style.fontFeatureSettings, "\"vert\" 2");
  }, "sibling-index() should not be allowed in @font-face descriptors");

  test(() => {
    assert_equals(font_face_rules[1].style.fontWeight, "300");
    assert_equals(font_face_rules[1].style.fontFeatureSettings, "\"vert\" 2");
  }, "sibling-count() should not be allowed in @font-face descriptors");

  const font_palette_rules = font_palette_sheet.sheet.cssRules;

  test(() => {
    assert_equals(font_palette_rules[0].basePalette, "1");
    assert_equals(font_palette_rules[0].overrideColors, "1 green");
  }, "sibling-index() should not be allowed in @font-palette-values descriptors");

  test(() => {
    assert_equals(font_palette_rules[1].basePalette, "1");
    assert_equals(font_palette_rules[1].overrideColors, "1 green");
  }, "sibling-count() should not be allowed in @font-palette-values descriptors");
</script>
