third_party/blink/web_tests/external/wpt/css/css-values/tree-counting/sibling-function-descriptors.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] sibling-index() should not be allowed in @page properties
  assert_equals: expected "100px" but got "calc(0px * sibling-index())"
[FAIL] sibling-count() should not be allowed in @page properties
  assert_equals: expected "100px" but got "calc(0px * sibling-count())"
Harness: the test ran to completion.
